pyproject.toml
[tool.poetry]
name = "runpod-ssh"
version = "0.1.0"
description = "RunPod SSH config manager"
authors = ["Avram Djordjevic <avramdjordjevic2@gmail.com>"]
readme = "README.md"
packages = [{include = "runpod_ssh"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
click = "^8.1.7"
pydantic = "^2.6.1"
platformdirs = "^4.2.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
runpod-ssh = "runpod_ssh.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
